Modified build-script to sign all artifacts using GPG before uploading them to the Nexus repo.
Very Important!
The build is bugged and running gradle uploadArchives twice is needed to get the signed pom uploaded as well. This should be fixed until the next release and the signing should be rewritten using huxi's pgp plugin for gradle. The plugin is a work in progress and is available here: https://github.com/huxi/huxi-gradle-plugins